+++ frontend/eda/useEdaNavigation.tsx
@@ -3,6 +3,7 @@ import { useTranslation } from 'react-i18next';
 import { Navigate } from 'react-router-dom';
 import { PageNavigationItem } from '../../framework/PageNavigation/PageNavigationItem';
 import { Login } from '../common/Login';
+import { EdaRoute } from './EdaRoutes';
 import { CredentialDetails } from './Resources/credentials/CredentialDetails';
 import { Credentials } from './Resources/credentials/Credentials';
 import { CreateCredential, EditCredential } from './Resources/credentials/EditCredential';
@@ -37,67 +38,6 @@ import { RuleDetails } from './rules/RuleDetails';
 import { Rules } from './rules/Rules';
 import { RuleAudit } from './views/RuleAudit/RuleAudit';
 import { RuleAuditDetails } from './views/RuleAudit/RuleAuditDetails';
-
-export enum EdaRoute {
-  Eda = 'eda',
-
-  Dashboard = 'eda-dashboard',
-
-  RuleAudits = 'eda-rule-audits',
-  RuleAuditDetails = 'eda-rule-audit-details',
-
-  RulebookActivations = 'eda-rulebook-activations',
-  CreateRulebookActivation = 'eda-create-rulebook-activation',
-  RulebookActivationDetails = 'eda-rulebook-activation-details',
-  RulebookActivationHistory = 'eda-rulebook-activation-history',
-  RulebookActivationInstancesDetails = 'eda-rulebook-activation-instances-details',
-
-  Rulebooks = 'eda-rulebooks',
-  RulebookDetails = 'eda-rulebook-details',
-
-  Projects = 'eda-projects',
-  CreateProject = 'eda-create-project',
-  EditProject = 'eda-edit-project',
-  ProjectDetails = 'eda-project-details',
-
-  DecisionEnvironments = 'eda-decision-environments',
-  CreateDecisionEnvironment = 'eda-create-decision-environment',
-  EditDecisionEnvironment = 'eda-edit-decision-environment',
-  DecisionEnvironmentDetails = 'eda-decision-environment-details',
-
-  Credentials = 'eda-credentials',
-  CreateCredential = 'eda-create-credential',
-  EditCredential = 'eda-edit-credential',
-  CredentialDetails = 'eda-credential-details',
-
-  Groups = 'eda-groups',
-  CreateGroup = 'eda-create-group',
-  EditGroup = 'eda-edit-group',
-  GroupDetails = 'eda-group-details',
-
-  Users = 'eda-users',
-  CreateUser = 'eda-create-user',
-  EditUser = 'eda-edit-user',
-  UserDetails = 'eda-user-details',
-  UserTokens = 'eda-user-tokens',
-
-  Rules = 'eda-rules',
-  CreateRule = 'eda-create-rule',
-  EditRule = 'eda-edit-rule',
-  RuleDetails = 'eda-rule-details',
-
-  Roles = 'eda-roles',
-  CreateRole = 'eda-create-role',
-  EditRole = 'eda-edit-role',
-  RoleDetails = 'eda-role-details',
-
-  EdaMyDetails = 'eda-my-details',
-  EdaMyTokens = 'eda-my-tokens',
-
-  CreateControllerToken = 'eda-create-controller-token',
-
-  Login = 'eda-login',
-}
 
 export function useEdaNavigation() {
   const { t } = useTranslation();
